+++ aeron-client/src/test/java/uk/co/real_logic/aeron/PublicationTest.java
@@ -26,8 +26,8 @@ import static uk.co.real_logic.aeron.common.concurrent.logbuffer.LogBufferDescri
 public class PublicationTest
 {
     public static final String DESTINATION = "udp://localhost:40124";
-    public static final long CHANNEL_ID_1 = 2L;
-    public static final long SESSION_ID_1 = 13L;
+    public static final int CHANNEL_ID_1 = 2;
+    public static final int SESSION_ID_1 = 13;
     public static final int TERM_ID_1 = 1;
     public static final int SEND_BUFFER_CAPACITY = 1024;
 
@@ -144,5 +144,4 @@ public class PublicationTest
             verify(buffer, times).close();
         }
     }
-
 }
